% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-commits.R
\name{compare_commit}
\alias{compare_commit}
\alias{compare_branch}
\title{Directory comparison between commits}
\usage{
compare_commit(commit_1 = "HEAD", commit_2 = "")

compare_branch(branch_1 = "master", branch_2 = "")
}
\arguments{
\item{commit_1}{Name (sha1, tag) of the origin commit to compare to (left)}

\item{commit_2}{Name (sha1, tag) of the destination commit to compare (right)}

\item{branch_1}{Name of the origin branch to compare to (left)}

\item{branch_2}{Name of the destination branch to compare (right)}
}
\description{
Compare directory to itself in another commit
}
\details{
For instance, use "" for the current state, "HEAD" for the last commit or
"master" to compare to branch master.

Add new files to git if you want to see them in the comparison
(with git add <file> or Staged in RStudio git pane)
}
