% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-git.R
\name{compare_commit_interactive}
\alias{compare_commit_interactive}
\title{Compare commits with interactive choice}
\usage{
compare_commit_interactive(ask_right = TRUE)
}
\arguments{
\item{ask_right}{Logical. Whether to ask for a second commit to compare.
FALSE will compare to current state.}
}
\description{
Compare commits with interactive choice
}
