% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_meld}
\alias{compare_meld}
\title{Compare meld}
\usage{
compare_meld(file_1, file_2 = NULL)
}
\arguments{
\item{file_1}{First file to compare.}

\item{file_2}{Second file to compare.}
}
\description{
Compare meld
}
\examples{
\dontrun{compare_meld()}
}
