% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_project_with_repo}
\alias{compare_project_with_repo}
\title{Compare with repo (project)}
\usage{
compare_project_with_repo()
}
\description{
Compare with repo (project)
}
\examples{
\dontrun{compare_project_with_repo()}
}
