% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_with}
\alias{compare_with}
\title{Version-control comparison of files}
\usage{
compare_with()
}
\value{
Invisibly returns the result of the meld command.
}
\description{
Version-control comparison of files
}
\examples{
\dontrun{compare_with()}
}
