% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_with_neighbor}
\alias{compare_with_neighbor}
\title{Compare with neighbor}
\usage{
compare_with_neighbor()
}
\description{
Compare active file with other file (search same directory).
}
\examples{
\dontrun{compare_with_neighbor()}
}
