% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_with_repo}
\alias{compare_with_repo}
\title{Compare with repo}
\usage{
compare_with_repo()
}
\description{
Compare with repo
}
\details{
\code{compare_with_repo} and \code{compare_project_with_repo} compare the
current file and RStudio project with the version control repository
version.
}
\examples{
\dontrun{compare_with_repo()}
}
