% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{compareWith-addins}
\alias{compareWith-addins}
\title{Comparison addins}
\description{
RStudio \href{https://rstudio.github.io/rstudioaddins/}{addins} for comparing
active files or projects against local or version control repository items.
}
\section{Addins}{

\describe{
\item{\strong{Compare with repo}}{Compare active file with version control repository.}
\item{\strong{Compare with repo - project}}{Compare active project with version control repository.}
\item{\strong{Compare with other...}}{Compare active file with another.}
\item{\strong{Compare with Git...}}{Compare active file with an arbitrary Git revision.}
\item{\strong{Compare with Git... - project}}{Compare active project with an arbitrary Git revision.}
\item{\strong{Compare Git revisions... - project}}{Compare changes between arbitrary Git revisions for the active project.}
}
}

\seealso{
Functions \code{\link[=compare_with_other]{compare_with_other()}} and \code{\link[=compare_with_repo]{compare_with_repo()}} for the flexible
comparison of files and directories.

Functions \code{\link[=compare_with_git]{compare_with_git()}} and \code{\link[=compare_git_revisions]{compare_git_revisions()}} for the flexible
comparison against or between arbitrary Git revisions.
}
