% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_with}
\alias{compare_with}
\alias{compare_with_neighbor}
\alias{compare_with_repo}
\alias{compare_project_with_repo}
\title{Version-control comparison of files}
\usage{
compare_with()

compare_with_neighbor()

compare_with_repo()

compare_project_with_repo()
}
\value{
Invisibly returns the result of calling the \code{meld} command via
\code{\link[=system2]{system2()}}.
}
\description{
Compare active files or projects against local or version control repository
items. Functions are called by the \pkg{compareWith} addins.
}
\section{Functions}{
\itemize{
\item \code{compare_with}: compares active file with another (search working directory). Serves as binding for addin 'Compare with...'.

\item \code{compare_with_neighbor}: compares active file with another (search same directory). Serves as binding for addin 'Compare with neighbor...'.

\item \code{compare_with_repo}: compares active file with the version control repository. Serves as binding for addin 'Compare with repo'.

\item \code{compare_project_with_repo}: compares active RStudio project with the version control repository. Serves as binding for addin 'Compare with repo (project)'.
}}

\examples{
\dontrun{compare_with()}
\dontrun{compare_with_neighbor()}
\dontrun{compare_with_repo()}
\dontrun{compare_project_with_repo()}
}
