% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_with}
\alias{compare_with}
\alias{compare_with_neighbor}
\alias{compare_with_repo}
\alias{compare_project_with_repo}
\alias{compare_meld}
\title{Version-control comparison of files}
\usage{
compare_with()

compare_with_neighbor()

compare_with_repo()

compare_project_with_repo()

compare_meld(file_1, file_2 = NULL)
}
\arguments{
\item{file_1}{First file to compare.}

\item{file_2}{Second file to compare.}
}
\value{
Invisibly returns the result of the meld command.
}
\description{
Compares active file with another (search working directory). Functions are called using Addins.
}
\section{Functions}{
\itemize{
\item \code{compare_with}: compares active file with another (search working directory). Serves as binding for addin 'Compare with...'.

\item \code{compare_with_neighbor}: compares active file with another within the same directory. Serves as binding for addin 'Compare with neighbor...'.

\item \code{compare_with_repo}: compares the current file and RStudio project with the version control repository version. Serves as binding for addin 'Compare with repo'.

\item \code{compare_project_with_repo}: compares the current file and RStudio project with the version control repository version. Serves as binding for addin 'Compare with repo (project)'.

\item \code{compare_meld}: calls meld.
}}

\examples{
\dontrun{compare_with()}
\dontrun{compare_with_neighbor()}
\dontrun{compare_with_repo()}
\dontrun{compare_project_with_repo()}
\dontrun{compare_meld()}
}
