% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareWith-addins}
\alias{compareWith-addins}
\alias{compare_with_other}
\alias{compare_with_repo}
\alias{compare_project_with_repo}
\title{Comparison addins}
\usage{
compare_with_other()

compare_with_repo()

compare_project_with_repo()
}
\value{
Invisibly returns the result of calling \code{meld} via \code{\link[=system2]{system2()}}.
}
\description{
Compare active files or projects against local or version control repository
items. Functions are used as bindings for the \pkg{compareWith} RStudio
addins.
}
\section{Functions}{
\itemize{
\item \code{compare_with_other}: compares active file with another. Serves as binding for addin 'Compare with other...'.

\item \code{compare_with_repo}: compares active file with the version control repository. Serves as binding for addin 'Compare with repo'.

\item \code{compare_project_with_repo}: compares active RStudio project with the version control repository. Serves as binding for addin 'Compare with repo (project)'.
}}

\examples{
\dontrun{compare_with_other()}

\dontrun{compare_with_repo()}

\dontrun{compare_project_with_repo()}

}
\seealso{
Function \code{\link[=compare_with]{compare_with()}} for the flexible comparison of files.
}
