% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareWith-addins}
\alias{compareWith-addins}
\title{Comparison addins}
\description{
RStudio \href{https://rstudio.github.io/rstudioaddins/}{addins} for comparing
active files or projects against local or version control repository items.
}
\section{Addins}{

\describe{
\item{\strong{Compare with repo}}{Compare active file with version control repository.}
\item{\strong{Compare with repo - project}}{Compare active project with version control repository.}
\item{\strong{Compare with other...}}{Compare active file with another.}
}
}

\seealso{
Functions \code{\link[=compare_with_other]{compare_with_other()}} and \code{\link[=compare_with_repo]{compare_with_repo()}} for the
flexible comparison of files and directories.
}
